Move reimbursement serializer to async task
