+++ jarbas/core/tests/test_reimbursements_command.py
@@ -14,81 +14,6 @@ class TestCommand(TestCase):
         self.command = Command()
 
 
-class TestSerializer(TestCommand):
-
-    def test_serializer(self):
-        expected = {
-            'applicant_id': 13,
-            'batch_number': 9,
-            'cnpj_cpf': '11111111111111',
-            'congressperson_document': 2,
-            'congressperson_id': 1,
-            'congressperson_name': 'Roger That',
-            'document_id': 42,
-            'document_number': '6',
-            'document_type': 7,
-            'document_value': 8.90,
-            'installment': 7,
-            'issue_date': '1970-01-01',
-            'leg_of_the_trip': '8',
-            'month': 1,
-            'net_values': '1.99,2.99',
-            'party': 'Partido',
-            'passenger': 'John Doe',
-            'reimbursement_numbers': '10,11',
-            'reimbursement_values': '12.13,14.15',
-            'remark_value': 1.23,
-            'state': 'UF',
-            'subquota_description': 'Subquota description',
-            'subquota_group_description': 'Subquota group desc',
-            'subquota_group_id': 5,
-            'subquota_id': 4,
-            'supplier': 'Acme',
-            'term': 1970,
-            'term_id': 3,
-            'total_net_value': 4.56,
-            'total_reimbursement_value': None,
-            'year': 1970,
-            'probability': None,
-            'suspicions': None
-        }
-        input = {
-            'applicant_id': '13',
-            'batch_number': '9',
-            'cnpj_cpf': '11111111111111',
-            'congressperson_document': '2',
-            'congressperson_id': '1',
-            'congressperson_name': 'Roger That',
-            'document_id': '42',
-            'document_number': '6',
-            'document_type': '7',
-            'document_value': '8.90',
-            'installment': '7',
-            'issue_date': '01/01/1970',
-            'leg_of_the_trip': '8',
-            'month': '1',
-            'net_values': '1.99,2.99',
-            'party': 'Partido',
-            'passenger': 'John Doe',
-            'reimbursement_numbers': '10,11',
-            'reimbursement_values': '12.13,14.15',
-            'remark_value': '1.23',
-            'state': 'UF',
-            'subquota_description': 'Subquota description',
-            'subquota_group_description': 'Subquota group desc',
-            'subquota_group_id': '5',
-            'subquota_number': '4',
-            'supplier': 'Acme',
-            'term': '1970',
-            'term_id': '3',
-            'total_net_value': '4.56',
-            'reimbursement_value_total': 'NaN',
-            'year': '1970'
-        }
-        self.maxDiff = 2 ** 10
-        self.assertEqual(self.command.serialize(input), expected)
-
-
 class TestCreate(TestCommand):
 
     @patch.object(Command, 'print_count')
@@ -140,10 +65,9 @@ class TestFileLoader(TestCommand):
 
     @patch('jarbas.core.management.commands.reimbursements.lzma')
     @patch('jarbas.core.management.commands.reimbursements.csv.DictReader')
-    @patch('jarbas.core.management.commands.reimbursements.Command.serialize')
-    def test_reimbursement_property(self, serializer, row, lzma):
+    def test_reimbursement_property(self, row, lzma):
         lzma.return_value = StringIO()
         row.return_value = dict(ahoy=42)
         self.command.path = 'reimbursements.xz'
-        list(self.command.reimbursements)
-        self.assertEqual(1, serializer.call_count)
+        reimbursements = tuple(self.command.reimbursements)
+        self.assertEqual(1, len(reimbursements))
